+++ src/main/java/com/esad/emasys/controller/AuthController.java
@@ -3,7 +3,7 @@ package com.esad.emasys.controller;
 import com.esad.emasys.model.Employee;
 import com.esad.emasys.model.LoginRequest;
 import com.esad.emasys.model.LoginResponse;
-import com.esad.emasys.services.interfaces.AuthService;
+import com.esad.emasys.service.interfaces.AuthService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
