+++ src/main/java/com/esad/emasys/controller/EmployeeController.java
@@ -3,9 +3,9 @@ package com.esad.emasys.controller;
 import com.esad.emasys.model.Department;
 import com.esad.emasys.model.Employee;
 import com.esad.emasys.model.Position;
-import com.esad.emasys.services.interfaces.DepartmentService;
-import com.esad.emasys.services.impl.EmployeeServiceImpl;
-import com.esad.emasys.services.interfaces.PositionService;
+import com.esad.emasys.service.interfaces.DepartmentService;
+import com.esad.emasys.service.impl.EmployeeServiceImpl;
+import com.esad.emasys.service.interfaces.PositionService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
