+++ src/main/java/com/esad/emasys/service/impl/LeaveServiceImpl.java
@@ -1,9 +1,9 @@
-package com.esad.emasys.services.impl;
+package com.esad.emasys.service.impl;
 
 import com.esad.emasys.model.Employee;
 import com.esad.emasys.model.Leave;
 import com.esad.emasys.repository.LeaveRepository;
-import com.esad.emasys.services.interfaces.LeaveService;
+import com.esad.emasys.service.interfaces.LeaveService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
