+++ src/main/java/com/esad/emasys/service/interfaces/DepartmentService.java
@@ -1,4 +1,4 @@
-package com.esad.emasys.services.interfaces;
+package com.esad.emasys.service.interfaces;
 
 import com.esad.emasys.model.Department;
 import com.esad.emasys.repository.DepartmentRepository;
